requirements.txt
numpy>=1.24.0
Pillow>=10.0.0
litellm>=1.30.0
typing>=3.7.4 